requirements.txt
huggingface_hub==0.33.0
nltk==3.9.1
pandas==2.2.3
psycopg2-binary==2.9.10
sacremoses==0.1.1
sentencepiece==0.2.0
streamlit==1.45.1
streamlit-aggrid==1.1.5.post1
dotenv==0.9.9
bcrypt==4.3.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
streamlit-modal==0.1.2
bert-score==0.3.13
